// SPA/caams-portal/src/app/auth/asset-subcategory/model/asset-subcategory.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AssetSubCategory { 
    assetCategoryName: string;
    assetCategoryOid: string;
    subCategoryName: string;
    assetSubCategoryOid: string;
    assetSubCategorySpecificationJson?: string;
    assetSubCategoryTypeOid?: string;
    createdBy: string;
    createdOn: Date;
    id?: number;
    organizationOid: string;
    updatedBy?: string;
    updatedOn?: Date;
}